Change interface and names
