+++ index.js
@@ -1,4 +1,4 @@
 'use strict'
-const IdentityProvider = require('./src/identity-providers')
+const IdentityProvider = require('./src/identity-provider')
 
 module.exports = IdentityProvider
